// console/src/api/generated/models/moment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.17.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Metadata } from './metadata';
// May contain unused imports in some cases
// @ts-ignore
import type { MomentSpec } from './moment-spec';
// May contain unused imports in some cases
// @ts-ignore
import type { MomentStatus } from './moment-status';

/**
 * 
 * @export
 * @interface Moment
 */
export interface Moment {
    /**
     * 
     * @type {string}
     * @memberof Moment
     */
    'apiVersion': string;
    /**
     * 
     * @type {string}
     * @memberof Moment
     */
    'kind': string;
    /**
     * 
     * @type {Metadata}
     * @memberof Moment
     */
    'metadata': Metadata;
    /**
     * 
     * @type {MomentSpec}
     * @memberof Moment
     */
    'spec': MomentSpec;
    /**
     * 
     * @type {MomentStatus}
     * @memberof Moment
     */
    'status'?: MomentStatus;
}